FSM구조 완성 / 적의 Roaming추가
FSM구조를 완성해서 이제 살만 붙이면
적 움직이는 거 다 만들 수 있음

+++ Assets/2.Scripts/Unit/UnitMove.cs
@@ -9,7 +9,7 @@ public class UnitMove : MonoBehaviour
     public Vector2 MoveDir { get { return moveDir; } }
     
     private Rigidbody2D rb2D;
-    [SerializeField] private int speed = 5;
+    [SerializeField] private float speed = 5;
 
     public UnityEvent<float> OnVelocityChange;
     private void Awake()
